http2: split tests of  http2.js into multiple files
